Add scoring for attendance

+++ src/main/java/ru/ifmo/neerc/volunteers/entity/Event.java
@@ -1,23 +1,14 @@
 package ru.ifmo.neerc.volunteers.entity;
 
-import java.util.HashSet;
-import java.util.Set;
-
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.OrderBy;
-import javax.validation.constraints.NotNull;
-
-import org.hibernate.validator.constraints.NotEmpty;
-
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
+import org.hibernate.validator.constraints.NotEmpty;
+
+import javax.persistence.*;
+import javax.validation.constraints.NotNull;
+import java.util.HashSet;
+import java.util.Set;
 
 /**
  * Created by Lapenok Akesej on 25.02.2017.
@@ -44,7 +35,7 @@ public class Event {
     private String information;
 
     @NotNull
-    private int attendanceValue;
+    private double attendanceValue;
 
     @OneToMany(fetch = FetchType.LAZY, mappedBy = "event")
     @OrderBy("id ASC")
